+++ cmd/smyrna/viewport.c
@@ -345,8 +345,8 @@ void init_viewport(ViewInfo * view)
     view->Topview->parms.hier.dist2_limit = 1;
     view->Topview->parms.hier.min_nvtxs = 20;
     view->Topview->parms.repos.rescale = Polar;
-    view->Topview->parms.repos.width = view->bdxRight-view->bdxLeft;
-    view->Topview->parms.repos.height = view->bdyTop-view->bdyBottom;
+    view->Topview->parms.repos.width =(int) (view->bdxRight-view->bdxLeft);
+    view->Topview->parms.repos.height =(int) (view->bdyTop-view->bdyBottom);
     view->Topview->parms.repos.margin = 0;
     view->Topview->parms.repos.graphSize = 100;
     view->Topview->parms.repos.distortion = 1.0;
